Avoid race conditions setting baselayer on location editor
Fixes a race condition that lead to using the wrong base layer on
location editor (only when navigating from another page, not on a
reload of the post create/edit page). We fix this by delaying render
of the leaflet-directive till map settings are loaded.

I think the cause of the error was that $scope.layers was populated,
then layers were bound into the leaflet directive, then we updated
$scope.layers *after* that and the changes didn't stick.

Fixes ushahidi/platform#1302