Added create rights filter to add-post-button directive
Added a lookup in the form roles resource to determine if the current
user has access to add to a survey. If the answer is no, that survey is
removed from the list of surveys in the add-post-button menu.

I’d much rather do this by adding a can-add field to the response
payload of the Forms end-point.

ushahidi/platform#1282