Guidelines for internal testing
This is a simple doc to help us improve how we do testing in ushahidi. While we don't want to pollute the PR or ISSUE templates with more guidelines, it's helpful to know what constitutes a good PR checklist and how to make one. We can iterate on this before it goes into develop.